backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0

# AI
together==1.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1
starlette==0.27.0

# Optional: Memory/retrieval
faiss-cpu==1.8.0
sentence-transformers==2.7.0
